Adding a redundancy filter that removes the subgroups that are too similar to others.
